+++ ICD.Connect.Settings/Attributes/SettingsProperties/AbstractSettingsPropertyAttribute.cs
@@ -4,7 +4,7 @@ using ICD.Common.Utils.Attributes;
 namespace ICD.Connect.Settings.Attributes.SettingsProperties
 {
 	[AttributeUsage(AttributeTargets.Property)]
-	public abstract class AbstractSettingsPropertyAttribute : AbstractIcdAttribute
+	public abstract class AbstractSettingsPropertyAttribute : AbstractIcdAttribute, ISettingsPropertyAttribute
 	{
 	}
 }